change restart on abnormal
It happened that the Tapo smart plugs got automatically updated firmware which caused the tapo Python package no longer work and upon querying the device properties, it ended with exception due to missing key. This got fixed in newer release of tapo. Anyhow, this caused the service to restart incessantly to no effect other than burning the CPU and causing memory churn. Restart set to always is not really helfpful for this service.